[core] chore: rename file sectionPanel.tsx
