+++ packages/core/src/components/index.ts
@@ -98,7 +98,7 @@ export { HandleHtmlProps, HandleInteractionKind, HandleProps, HandleType } from 
 export { MultiSlider, MultiSliderProps, SliderBaseProps } from "./slider/multiSlider";
 export { NumberRange, RangeSlider, RangeSliderProps } from "./slider/rangeSlider";
 export { Section, SectionElevation, SectionProps } from "./section/section";
-export { SectionCard, SectionCardProps } from "./section/sectionPanel";
+export { SectionCard, SectionCardProps } from "./section/sectionCard";
 export { Slider, SliderProps } from "./slider/slider";
 export { Spinner, SpinnerProps, SpinnerSize } from "./spinner/spinner";
 export { Tab, TabId, TabProps } from "./tabs/tab";
