+++ packages/core/src/components/section/sectionCard.tsx
